refactor: add modals folder @src/features/modals/ui/modalAddPhoto: modalAddPhoto.module.scss, modalAddPhoto.tsx, useModalAddPhoto.ts, index.ts
- Add in  _mixins image-preview and image-wrapper mixins
- Add index.ts files for modals and modalAddPhoto
