+++ src/app/styles/_mixins.scss
@@ -51,3 +51,38 @@
     color: var(--light-100);
   }
 }
+
+@mixin image-preview($size) {
+  position: relative;
+
+  overflow: hidden;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: $size;
+  height: $size;
+  margin: 0;
+  padding: 0;
+
+  background-color: var(--dark-500);
+}
+
+@mixin image-wrapper {
+  position: absolute;
+
+  overflow: hidden;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 100%;
+  height: 100%;
+
+  & > img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    object-position: center;
+  }
+}
