+++ src/features/index.ts
@@ -1,3 +1,4 @@
 export * from './auth'
 export * from './layout'
+export * from './modals'
 export * from './profile'
