+++ src/features/modals/ui/modalAddPhoto/index.ts
@@ -1,0 +1,1 @@
+export * from './modalAddPhoto'
